Fix model m2m error, migration
